WIP checkpoint: GPU M2 int64 path + [gpu] kill-switch semantics
HELD - M2 build awaits explicit owner go (per GPU_DESIGN.md M2 scope and
the audit-wave decision). Do NOT merge without approval.

This checkpoint starts the M2 design agreed with the owner:
- gpu_metal.m: buffer cache re-sized in BYTES (was floats) so the coming
  int64 exact-result path (MSL 'long' = 64-bit) shares the float32
  buffers; declared wyn_gpu_map_i64_begin/_run in the header contract.
- toml.h [gpu]: reframed 'enabled' as a global kill-switch defaulting to
  TRUE (incl. absent section/file), added enabled_set (key-present flag)
  and float_enabled (explicit opt-in for the f32 float precision
  contract). Matches the agreed model: int dispatch automatic, float
  behind documented-epsilon opt-in.

Builds clean. Not wired to the i64 codegen path yet.

+++ src/gpu_metal.m
@@ -20,6 +20,8 @@
 //   wyn_gpu_available()                  -> 1 if a Metal device exists
 //   wyn_gpu_map_f32_begin(n, &in_ptr)    -> 0 ok; caller packs into in_ptr
 //   wyn_gpu_map_f32_run(n, src, &out)    -> 0 ok; result readable at out
+//   wyn_gpu_map_i64_begin / _run         -> same shape, int64 elements (MSL
+//                                           `long` is 64-bit: exact results)
 //   (buffers are cached and reused across calls; nothing to free per call)
 
 #ifdef __APPLE__
@@ -40,9 +42,10 @@ static struct { const char* src; void* pso; } _gpu_psos[WYN_GPU_PSO_CACHE];
 static int _gpu_pso_count = 0;
 
 // Cached shared-storage buffers, grown as needed and reused across calls.
+// Sized in BYTES so the float32 and int64 paths share the same pair.
 static id<MTLBuffer> _gpu_in_buf = nil;
 static id<MTLBuffer> _gpu_out_buf = nil;
-static int _gpu_buf_cap = 0;      // capacity in floats
+static size_t _gpu_buf_cap = 0;   // capacity in bytes
 
 static int _gpu_init(void) {
     if (_gpu_init_state != 0) return _gpu_init_state;
